chore: move retry_with_mockable_time method in quickwit-common::retry
