+++ quickwit/quickwit-storage/src/object_storage/s3_compatible_storage.rs
@@ -36,8 +36,10 @@ use aws_smithy_http::byte_stream::ByteStream;
 use base64::prelude::{Engine, BASE64_STANDARD};
 use futures::{stream, StreamExt};
 use once_cell::sync::{Lazy, OnceCell};
+use quickwit_aws::error::RetryableWrapper;
 use quickwit_aws::get_aws_config;
-use quickwit_aws::retry::{retry, Retry, RetryParams, Retryable};
+use quickwit_common::retry::RetryParams;
+use quickwit_aws::retry::{retry, Retry};
 use quickwit_common::uri::Uri;
 use quickwit_common::{chunk_range, into_u64_range};
 use quickwit_config::S3StorageConfig;
